upgrade/requirements.txt

django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-filter
psycopg2-binary
celery
redis
